Eliot 1860 trip (template for other trips)
Grace, I have updated this trip and corrected the messed up coordinates. You should be able to use exactly this syntax and replace the coordinates with the longitude/latitude data from the CSV files to create three more trips for 1861, 1864, and 1869, and save them as .js files, declaring the variables as var = eliot_1861trip, etc. If you get that far but it is not obvious how to go from there (include the script, call the script as a layer), then just e-mail me and I'll have another look. Otherwise, I will just leave you to it.